moves table into module
